support secure_metadata on resume (#183)
* NinchatSession: delegate on deallocate

The delegate is called if the session had not been ended before. Thus, the delegate won’t be called twice ever.

* Debugger: show error only in debug mode

* UnitTest: Swift Tests: improve tests

Add a test for nested dictionaries. Fix a build issue with new WebRTC function.

* Extensions: UserDefaults: Add UserDefault

Add the extension with ‘save’, ‘load’, and ‘remove’ functions. Add tests for Swift and LowLevel types

* LowLevelClient Framework: Update to version 0.3.10

The new version includes new functions, `marshal` and `unmarshal`

* SessionManager: save/load metadata to/from storage

Instead of dealing with the injected object, the `metadata` is always read from the UserDefaults.

* Metadata: remove metadata on session start and end

* Podspec: point to latest version of LowLevelClient

* Podfile: point to latest version of LowLevelClient

* Questionnaire: Add audienceMetadata to answers

When the user sends `pre_answers` to the server, it has to be alogside metadata given by the user during the initialization (which includes `secure` ones)

* MarshalJSON: remove warning

Replacing `NSErrorPointer` with `&NSError`

* SessionImpl: sync changed metadata for pre_answers

* SessionMetadata: extract to dedicated protocol

Extract metadata updates into a dedicated protocol.

* SessionMetadata: fix issue with pre_answers

With an issue for setting given pre_answers during session initialization

* github-ci: fix tests