UserDefaults: migrate from ‘standard’ to ‘ninchat’ suite
When using `UserDefaults.standard`, the host app can read/write the metadata outside of the SDK. This can result in unexpected behaviours in some cases. With this commit, the SDK creates a separate suite to save and load metadata there.
The host app still **can override** the metadata, but it needs to explicitly access to the suite.