ViewModel: Questionnaire: make asnwers updateable
The ViewModel used to avoid any updates on submitted answers. This was
an approach to avoid running corresponded closures more than once.
However, the assumption is not true for the last item in the view, as it
is currently active and no closures can be run more than once. Thus, the
commit add a boolean to the function to let the caller specify if the
answers shall be updated or skipped.