Attachments: filter supported- unsupported formats
The commit filters unsupported formats and returns an error so the user
knows that the format is not usable (This is an exceptional case, as the
only problem is with slow-motion videos. Unfortunately, there is no
known workaround to filter those videos.)
The commit also fix the bug in sending videos with .jpg file extensions.
It also unifies photo extensions to '.jpg' to avoid confusion.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>