bump to version 0.4.8 - hotfix (#289)
* github-ci: improve jobs

The commit includes the following improvements:
    - Removing an extra task for building the SDK. Running 'Tests' tasks
      includes building the project before running tests' cases.
    - Removing an extra task for cleaning the SDK. While there has been
      no previous build on the machine, this task was completely useless.
    - Remove running extra jobs for pull requests, while both branches
      will be (or have been already) tested.
    - Switch from Xcode 12.2 to 12.4
    - Switch iOS Simulator from iPhone 8 to iPhone 11.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* InitialViewController: fix issue on rotating device

Rotating the device no longer messes up with the overridden componenets.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* QueueViewController: fix issue on rotating device

Rotating device no longer messes up with overridden componenets.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* RatingViewController: fix issue on rotating device

Rotating the device no longer messes up with overridden compoenents.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* SiteConfiguration: improve lookup algorithm

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* SiteConfiguration: fix issue in getting questionnaire style

The issue was caused due to unwanted optional variable.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Remove warnings for 'class' conformance in protocols

By replacing ': class' with ': AnyObject'

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Add/Update missing new override keys

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Remove unused variable warnings

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Improve overriding layer

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Coordinator: remove unused flag variables

Variables defined for deallocation purposes, but never have been used.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Coordinator: refactor and improve code style

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* InitialViewController: fix overriding view

The commit fixes an issue in overriding a view component that caused by
a missing 'else' statement

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Coordinator: improve performance on processing questionnaires

The problem started to appear after using the new overriding approach,
which apperantly uses more resources than legacy one. The combination
of the newly added resource usage with processing heavy questionnaires
resulted in an incomplete overriding view.
The problem is fixed by replacing simple 'DispatchQueues' with a more
advance 'OperationQueue' algorithm, and adding appropriate dependencies
between tasks in order to control devices' resources.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* queues/avoid duplicated (#276)

* queues/lazy describe (#277)

* Dictionary+Extension: recursively find values

Suitable for nested dictionaries.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* SiteConfiguration: refactor questionnaire getters

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* SessionManager: unify functions to describe queues

While 'list(queues:completion)' and 'describe(queuesID:completion)'
functions are not duplicated, but they are serving the same purpose.
Thus, they are unified by this commit

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* NINChatSession: improve describing all queues

The commit improves how all queues (that are accessible either from the
site_config, or directly injected by the host app) are described. This
commit, hopefully, will address reported inconsistencies in redirecting
to some queues.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* questionnaire/fix override (#278)

* QuestionnaireCell: override color for bot name

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Questionnaire: Select: fix overriding background

The background for Selected and Unselected states were applied in a
wrong order

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* DataSource: Questionnaire: make element selectable on back

Fixes a problem with re-selecting elements when user taps on "back"
navigation button in "form" questionnaires.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* ChatViewController: hide call invitation on hangup (#280)

With the commit, the video call invitation dialouge will be hidden if
the agent stop calling the user.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* ChatViewController: reload before loading history (#279)

Finxes a bug for cases of queue transfer when the chat page was not reloaded
properly and resulted in a messed up conversation.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* Storyboard: improve rating landscape view (#281)

For both normal and conversation styles

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* RatingViewController: skip post questionnaire (#282)

If user "skips" giving a rating in Rating ViewController, the post audience
questionnaire view will be skipped as well.

Signed-off-by: Hassan Shahbazi <hassan@ninchat.com>

* questionnaire/fix registration (#283)

* Questionnaire: remove retain cycles, improve style

* Coordinator: fix questionnaire compelete and register

* memory/remove retain cycles (#284)

* ViewController: remove retain-cycle

* DataSource/Delegate: remove retain-cycle

* InternalDelegate: improve how the internal delegate works

Instead of creating a new object (which should hold an instance to
NinchatSession and is very likely to result in a retain-cycle), the internal
delegate is changed to be just a simple extension of NinchatSession.

* Managers: add missing captures for many cases

* ServiceManager: fix a known memory leakage for URLSession

* remove dead code

* change 'class' to 'struct' for models

* Podspec: update to version 0.4.7

* questionnaire/update text closed queue (#287)

* SiteConfiguration: rename 'audienceRegisteredClosedText'

The commit fixes a naming issue in audience registered closed text,
which was named as 'audienceCloseRegisteredText'.
This resulted in not showing appropriate messages when needed, and also
not registering questionnaire data.

* Questionnaire: show 'audienceRegisteredClosedText' if queue is closed

* questionnaire/fix register cached metadata (#288)

* Questionnaire: register secure metadata if cached

The commit fixes an issue about registering audience in a queue where
the secure metadata is required, and the SDK must loads them from the
cache. By removing injection of the metadata instance, the view model
now directly access the metadata from session manager.

* github-ci: fix failed tests

* Podspec: update to version 0.4.8