feature/title bar (#290)
* Button: rename to 'NINButton'

'Button', which is customized 'UIButton' component being used throughout
the project, uses SwiftUI 'Button' reserved name. This conflict results
in errors when SwiftUI framework is imported.

This commit resolves the issue by renaming 'Button' to 'NINButton'.

* rename 'weakSelf' to '`self`'

To keep the consistency, instead of using a name for capturing weak
instances of 'self', it makes more sense to use the 'self' naming
convention.

* Coordinator: improve code style

* Constants: introduce overriding key for titlebar

* Constants: add support for 'Close' key

* UIColor: add colors for titlebar

* NINLowLevelClient: add property for job title

* Session: add public Agent to session interface

A new variable with the type ChannelUser that holds the agent's
information. The variable is then used to show information on the
titlebar. Also, jobTitle field has been added to the ChannelUser model.

* SiteConfiguration: add property for hideTitlebar

* Coordinator: hide navigationbar if title is not hidden

* Titlebar: add required protocol and impl for titlebar

The commit doesn't implement 'addTitleBar' function as it needs to be
done either via SwiftUI or UIKit.

* NINQueueViewController: add titlebar

* NINQuestionnaireViewController: add titlebar

Also, the commit adds required changes in the Coordinator to handle
titlebar Close action.

* NINChatViewController: add titlebar

* NINRatingViewController: add titlebar

* storyboard: add titlebar container to ViewControllers

* Apply avatar rules

The commit applies required rules that must be followed for showing or
hiding avatars across the project, including the titlebar

* QuestionnaireTypingCell: rename to 'loading' cell

* Titlebar: extract titlebar shaping function

* Titlebar: SwiftUI implementation

This commit adds titlebar implementation in SwiftUI. However, since
customers are not yet ready to migrate to iOS 13, the implementation
might be used later. This is why files are not linked to the target and
won't be installed with 'pod install'.

* Titlebar: UIKit implementation

Add an implemention using UIKit (xib, UIView) to support earlier iOS
versions. The implementation will be deprecated once we upgrade
deployment target to iOS 13 and above.

* PostQuestionnaire: apply rules to show titlebar

* Avatar: fix issue in initialization

The commit fixes an issue caused by commit c9c903f1f2 that resulted in using
questionnaire info (avatar,name) everywhere.

* Titlebar: improve titlebar name

Use agentName first and if it is null, use user_attrs.name

* Storyboard: re-install form rating component

* Titlebar: improve size. fix avatar setting issue

Fix an early return for setting titlebar's avatar

* RatingViewController: improve style

* github-ci: fix failed tests