Questionnaire: Cell: add more containers to the cell
In order to separate between the text in the bubble and options below
that, the commit introduces HasTitle and HasOptions protocols which
allow the corresponded cell to add each one into a different UIView.
The title and the buuble are in a same view that are embedded in a StackView
with the options view. This will let the cell to automatically update
view when either one is hidden.