Questionnaire: adopt to the new approach for closing the session
In the new approach, the questionnaire answers are saved first, then the guest user is deleted and after that, the session is closed. But, according to the previous commit, the SDK stays open and let the _completed element be shown. When the user taps on the navigation button of the exit element, the SDK is closed.
The new approach fixes some edge cases where a user is able to resume a finished session by not selecting the exit element.